@@ -1,10 +1,10 @@
 android {
-    compileSdkVersion 27
-    buildToolsVersion '27.0.3'
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
 
     defaultConfig {
         minSdkVersion 21
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode versionCode
         versionName version
         manifestPlaceholders = ['appAuthRedirectScheme': 'org.radarcns.android']
@@ -43,7 +43,7 @@ ext.description = 'RADAR Android OAuth2 LoginManager.'
 
 dependencies {
     api project(':radar-commons-android')
-    api 'net.openid:appauth:0.6.1'
+    api 'net.openid:appauth:0.7.1'
 }
 
 apply from: '../gradle/publishing.gradle'
