@@ -1,10 +1,10 @@
 android {
-    compileSdkVersion 27
-    buildToolsVersion '27.0.3'
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
 
     defaultConfig {
         minSdkVersion 21
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode versionCode
         versionName version
     }
@@ -41,7 +41,7 @@ ext.description = 'RADAR Android QR LoginManager.'
 
 dependencies {
     api project(':radar-commons-android')
-    implementation 'com.journeyapps:zxing-android-embedded:3.5.0'
+    implementation 'com.journeyapps:zxing-android-embedded:3.6.0'
 }
 
 apply from: '../gradle/publishing.gradle'
