@@ -15,12 +15,12 @@
  */
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion '27.0.3'
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
 
     defaultConfig {
         minSdkVersion 21
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode versionCode
         versionName version
     }
@@ -61,9 +61,9 @@ dependencies {
     }
     api 'org.radarcns:radar-schemas-commons:0.4.0'
 
-    implementation 'com.android.support:support-v4:27.1.1'
+    implementation 'com.android.support:support-v4:28.0.0'
 
-    implementation 'com.google.firebase:firebase-config:16.0.0'
+    implementation 'com.google.firebase:firebase-config:16.0.1'
     implementation 'com.crashlytics.sdk.android:crashlytics:2.9.5'
     implementation 'sk.baka.slf4j:slf4j-handroid:1.7.27'
 }
