@@ -167,7 +167,7 @@ public abstract class DeviceService<T extends BaseDeviceState> extends Service i
     private void doBind(Intent intent, boolean firstBind) {
         logger.info("Received (re)bind in {}", this);
         Bundle extras = BundleSerialization.getPersistentExtras(intent, this);
-        onInvocation(extras);
+        onInvocation(extras != null ? extras : new Bundle());
 
         RadarApplication application = (RadarApplication)getApplicationContext();
         application.onDeviceServiceInvocation(this, extras, firstBind);
@@ -512,8 +512,8 @@ public abstract class DeviceService<T extends BaseDeviceState> extends Service i
         }
         if (updatedSource == null) {
             // try again in a minute
-            handler.postDelayed(() -> stopDeviceManager(unsetDeviceManager()),
-                    ThreadLocalRandom.current().nextLong(1_000L, 120_000L));
+            long delay = ThreadLocalRandom.current().nextLong(1_000L, 120_000L);
+            handler.postDelayed(() -> stopDeviceManager(unsetDeviceManager()), delay);
             return;
         }
         AppAuthState auth = AppAuthState.Builder.from(result).build();
